User authentication and button authentication

+++ resources/views/posts/show.blade.php
@@ -7,9 +7,12 @@
     <hr>
     <small>written on {{$post->created_at}}</small>
     <hr>
-    <a href="/posts/{{$post->id}}/edit" class="btn btn-primary">Edit</a>
-
-    {!! Form::open(['action' => ['PostsController@destroy', $post->id], 'method' => 'DELETE', 'class' => 'float-right']) !!}
-        {{ Form::submit('Delete', ['class' => 'btn btn-danger']) }}
-    {!! Form::close() !!}
+    @if (!Auth::guest())
+        @if (Auth::user()->id == $post->user_id)
+            <a href="/posts/{{$post->id}}/edit" class="btn btn-primary">Edit</a>
+            {!! Form::open(['action' => ['PostsController@destroy', $post->id], 'method' => 'DELETE', 'class' => 'float-right']) !!}
+                {{ Form::submit('Delete', ['class' => 'btn btn-danger']) }}
+            {!! Form::close() !!}
+        @endif
+    @endif
 @endsection
